api/requirements.txt
orjson==3.9.10
fastapi==0.109.0
uvicorn==0.27.0
pulsefire==2.0.2
python-dotenv==1.0.0